+++ src/features/map/Map.js
@@ -5,8 +5,8 @@ import Menu from "../menu/Menu";
 import { ENDPOINT } from "./../requests";
 import ReactDOM from "react-dom";
 import styles from "./Map.module.css";
-import { isEqual } from "lodash";
-import SelectInput from "@material-ui/core/Select/SelectInput";
+import drawPath from './DrawPath';
+import removePath from './RemovePath';
 
 class Map extends React.Component {
   constructor(props) {
@@ -157,75 +157,6 @@ class Map extends React.Component {
     </div>
   );
 
-  removePath = (index, paths, matchExact = false) => {
-    const data = paths[index];
-
-    console.log("Remove path " + index)
-
-    const { lng: end_lng, lat: end_lat } = data.end_node;
-    const coords = data.coordinates;
-    console.log("Initial Map data: ")
-    const sourceData = this.map.getSource("lines")["_data"];
-    console.log(sourceData);
-    let newFeatures = [...this.map.getSource("lines")["_data"].features];
-    //Iterating in reverse so that modifying newFeatures while looping works
-
-    console.log("Pruning for: ");
-    console.log(data)
-
-    let spliceCount = 0;
-    for (let i = newFeatures.length - 1; i >= 0; i--) {
-      let featObjCoords = newFeatures[i]["geometry"]["coordinates"];
-      if (featObjCoords.length === 0) {
-        if (index === 0) {
-          newFeatures.splice(i, 1);
-          spliceCount++;
-        }
-      } else if (
-        isEqual(new Set(featObjCoords), new Set(coords)) ||
-        (end_lng === featObjCoords[0][0] && end_lat === featObjCoords[0][1] && !matchExact)
-      ) {
-        newFeatures.splice(i, 1);
-        spliceCount++;
-      }
-    }
-    console.log("SpliceCount: " + spliceCount);
-    this.map.getSource("lines").setData({
-      ...sourceData,
-      features: newFeatures,
-    });
-    console.log("Final Remove Map Data: ")
-    console.log(this.map.getSource("lines")["_data"]);
-  };
-
-  drawPath = (index, paths) => {
-    const data = paths[index];
-    const coords = data.coordinates;
-    const newLine = {
-      type: "Feature",
-      properties: {
-        // color: "#B5B5FE" // soft purple
-        color: "#9A21F9", // stronger purple
-      },
-      geometry: {
-        type: "LineString",
-        coordinates: coords,
-      },
-    };
-
-    const sourceData = this.map.getSource("lines")["_data"];
-    let newFeatures = [...sourceData.features];
-    newFeatures.push(newLine);
-
-    this.map.getSource("lines").setData({
-      ...sourceData,
-      features: newFeatures,
-    });
-
-    console.log("Final Draw Map Data: ")
-    console.log(this.map.getSource("lines")["_data"]);
-  };
-
   addPopup(nodeIndex) {
     const placeholder = document.createElement("div");
     const windowContent = this.markerDeletionWindow(nodeIndex);
@@ -278,7 +209,7 @@ class Map extends React.Component {
       newPaths.push(obj);
       this.setState(
         { paths: newPaths },
-        this.drawPath(this.state.paths.length - 1, this.state.paths)
+         drawPath(this.map, this.state.paths.length - 1, this.state.paths),
       );
     }
   }
@@ -339,16 +270,16 @@ class Map extends React.Component {
     if (new_paths.length === 0) {
       return;
     }
-    this.removePath(index, this.state.paths);
+    removePath(this.map, index, this.state.paths);
     new_paths.splice(index, 1);
     for (const [idx, value] of Object.entries(data)) {
-      this.removePath(idx, this.state.paths);
+      removePath(this.map, idx, this.state.paths);
       new_paths[idx] = value;
       this.setState({ paths: new_paths });
       if (new_paths.length === 0) {
         return;
       }
-      this.drawPath(idx, this.state.paths);
+      drawPath(this.map, idx, this.state.paths);
     }
     this.setState({
       paths: new_paths,
@@ -378,13 +309,13 @@ class Map extends React.Component {
           for (let [idx, value] of Object.entries(data["segment_updates"])) {
             if (parseInt(idx) > 0) {
               console.log("Removing path: " + idx)
-              this.removePath(idx, new_paths, true);
+              removePath(this.map, idx, new_paths, true);
               new_paths[idx] = value;
             }
           }
           for (let [idx, value] of Object.entries(data["segment_updates"])) {
             if (parseInt(idx) > 0) {
-              this.drawPath(idx, new_paths);
+              drawPath(this.map, idx, new_paths);
             }
           }
           markerCoordsCallback(data["new_node"].lng, data["new_node"].lat);
